+++ src/YAYL/YamlContext.cs
@@ -1,0 +1,7 @@
+namespace YAYL;
+
+public class YamlContext
+{
+    public string? FilePath {get; init;}
+    public string? WorkingDirectory {get; init;}
+}