+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.6.0] - 2025-05-11
+
+- **Adding YamlContext**: New configuration object that can be provided to parse operations.
+- **Adding YamlPathFieldAttribute**: New attribute that resolves file paths to absolute paths during parsing.
+
 ## [1.5.0] - 2025-05-09
 
 - **Parser support for Sets:** Adding support for parsing set types (`ISet<T>`, `HashSet<T>`, and `SortedSet<T>`)